cleaned up a few types, also fixing some messages #25
