+++ react-pokemon-client/src/utils/env.ts
@@ -1,11 +1,5 @@
-
-export const AppEnv = {
-  Development: 'development',
-  Preview: 'preview',
-  Production: 'production',
-} as const;
-
-export type AppEnv = (typeof AppEnv)[keyof typeof AppEnv];
+import {AppEnv as AppEnvLit} from '../../stagetypes'
+import type {AppEnv} from '../../stagetypes'
 
 function getEnvironmentName(): string | undefined {
     return process.env.VITE_APP_ENV || (typeof import.meta.env !== 'undefined' ? import.meta.env.VITE_APP_ENV : undefined);
@@ -13,22 +7,22 @@ function getEnvironmentName(): string | undefined {
 
 export function getAppEnv(): AppEnv{
     const env = getEnvironmentName();
-    if (env === AppEnv.Production) 
-        return AppEnv.Production;
-    if (env === AppEnv.Preview) 
-        return AppEnv.Preview;
+    if (env === AppEnvLit.Production) 
+        return AppEnvLit.Production;
+    if (env === AppEnvLit.Preview) 
+        return AppEnvLit.Preview;
     //the else case
-    return AppEnv.Development;
+    return AppEnvLit.Development;
 }
 
 export function isDev() {
-    return getAppEnv() === AppEnv.Development
+    return getAppEnv() === AppEnvLit.Development
 }
 
 export function isPreview(){
-    return getAppEnv() === AppEnv.Preview;
+    return getAppEnv() === AppEnvLit.Preview;
 }
 
 export function isProd() {
-    return getAppEnv() === AppEnv.Production;
+    return getAppEnv() === AppEnvLit.Production;
 }
